:unlock: Show QA helpers section on all dashboards.
Also add warning about how they should be removed before going to production.
