+++ web/src/components/dashboards/SystemAdminDashboard.vue
@@ -23,50 +23,11 @@
       <MySubscriptionsCard />
     </v-col>
   </v-row>
-  <v-divider
-    thickness="1"
-    class="my-8 border-opacity-100"
-  />
-  <v-row>
-    <v-col
-      cols="12"
-      md="6"
-    >
-      <QaScenariosCard />
-    </v-col>
-    <v-col
-      cols="12"
-      md="6"
-    >
-      <v-btn
-        class="mr-4"
-        color="primary"
-        variant="outlined"
-        :loading="isLoadingUserGroups"
-        @click="syncUserGroups"
-        >Sync User Groups</v-btn
-      >
-    </v-col>
-  </v-row>
+  <QaHelpersSection />
 </template>
 
 <script lang="ts" setup>
-import { ref } from "vue"
-
-import userGroupsApi from "@/api/user-groups-api"
-
 import DatasetSearchCard from "@/components/dashboard-widgets/DatasetSearchCard.vue"
 import MySubscriptionsCard from "@/components/dashboard-widgets/MySubscriptionsCard.vue"
-import QaScenariosCard from "@/components/qa-scenarios/QaScenariosCard.vue"
-
-const isLoadingUserGroups = ref(false)
-
-async function syncUserGroups() {
-  isLoadingUserGroups.value = true
-  try {
-    await userGroupsApi.sync()
-  } finally {
-    isLoadingUserGroups.value = false
-  }
-}
+import QaHelpersSection from "@/components/dashboard-widgets/QaHelpersSection.vue"
 </script>
